backend/microservices/auth-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
